Extract ListNavigationEnabled
